requirements.txt
absl-py==1.4.0
aioice==0.7.6
aiortc==1.4.0
altair==4.2.2
astunparse==1.6.3
attrs==22.2.0
av==10.0.0
blinker==1.5
blis==0.7.9
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
confection==0.0.4
contourpy==1.0.7
cryptography==39.0.2
cycler==0.11.0
cymem==2.0.7
decorator==5.1.1
dnspython==2.3.0
entrypoints==0.4
flatbuffers==23.3.3
fonttools==4.39.2
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-crc32c==1.5.0
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
idna==3.4
importlib-metadata==6.1.0
Jinja2==3.1.2
jsonschema==4.17.3
keras==2.11.0
keras-models==0.0.7
keras-utils==1.0.13
kiwisolver==1.4.4
langcodes==3.3.0
libclang==15.0.6.1
lxml==4.9.3
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
mediapipe==0.9.1.0
murmurhash==1.0.9
netifaces==0.11.0
numpy==1.24.2
oauthlib==3.2.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
pathlib==1.0.1
pathy==0.10.1
Pillow==9.4.0
preshed==3.0.8
protobuf==3.19.6
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.6
pydeck==0.8.0
pyee==9.0.4
Pygments==2.14.0
pylibsrtp==0.8.0
Pympler==1.0.1
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.3.2
rsa==4.9
semver==2.13.0
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
streamlit==1.20.0
streamlit-webrtc==0.45.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
thinc==8.1.9
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
wasabi==1.1.1
watchdog==3.0.0
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
